refactor(接口测试): 执行结果处理优化，嵌套事务拆分成小事务
--story=1010710 --user=赵勇 【bug转需求】优化消息通知机制（同步发1.20） https://www.tapd.cn/55049933/s/1315550
