+++ backend/pom.xml
@@ -535,7 +535,7 @@
                         <artifactItem>
                             <groupId>org.python</groupId>
                             <artifactId>jython-standalone</artifactId>
-                            <version>2.7.2</version>
+                            <version>2.7.0</version>
                             <type>jar</type>
                             <overWrite>true</overWrite>
                             <outputDirectory>src/main/resources/jmeter/lib/ext</outputDirectory>
